AshCubDB.Info (ash_cubdb v0.6.2)

Auto-generated introspection for the AshCubDB DSL.

Functions

cubdb_auto_compact?(dsl_or_extended)

@spec cubdb_auto_compact?(dsl_or_extended :: module() | map()) :: boolean()

Whether or not to automatically compact the CubDB database.

See the CubDB documentation for more information.

cubdb_auto_file_sync?(dsl_or_extended)

@spec cubdb_auto_file_sync?(dsl_or_extended :: module() | map()) :: boolean()

Whether or not to automatically flush the buffer to disk on write.

See the CubDB documentation

cubdb_directory(dsl_or_extended)

@spec cubdb_directory(dsl_or_extended :: module() | map()) ::
  {:ok, nil | String.t()} | :error

The directory within which to store the CubDB data.

If none is supplied, then one will be automatically generated in the priv directory of the parent OTP application.

cubdb_directory!(dsl_or_extended)

@spec cubdb_directory!(dsl_or_extended :: module() | map()) ::
  (nil | String.t()) | no_return()

The directory within which to store the CubDB data.

If none is supplied, then one will be automatically generated in the priv directory of the parent OTP application.

cubdb_name(dsl_or_extended)

@spec cubdb_name(dsl_or_extended :: module() | map()) :: {:ok, atom()} | :error

The name of the CubDB database.

By default this is the name of the resource module, however in some (rare) circumstances you may wish to specifically name the database.

cubdb_name!(dsl_or_extended)

@spec cubdb_name!(dsl_or_extended :: module() | map()) :: atom() | no_return()

The name of the CubDB database.

By default this is the name of the resource module, however in some (rare) circumstances you may wish to specifically name the database.

cubdb_options(dsl_or_extended)

@spec cubdb_options(dsl_or_extended :: module() | map()) :: %{
  required(atom()) => any()
}

cubdb DSL options

Returns a map containing the and any configured or default values.

cubdb_otp_app(dsl_or_extended)

@spec cubdb_otp_app(dsl_or_extended :: module() | map()) :: {:ok, atom()} | :error

The OTP application in whose priv directory data should be stored.

Only used if directory is not supplied. When not provided Application.get_application/1 will be called for the resource.

cubdb_otp_app!(dsl_or_extended)

@spec cubdb_otp_app!(dsl_or_extended :: module() | map()) :: atom() | no_return()

The OTP application in whose priv directory data should be stored.

Only used if directory is not supplied. When not provided Application.get_application/1 will be called for the resource.

field_layout(resource_or_dsl_state)

@spec field_layout(Ash.Resource.t() | Spark.Dsl.t()) :: nil | {tuple(), tuple()}

Retrieve the cached field layout for the resource.